app/requirements.txt
anyio
asyncpg
click
fastapi
greenlet
h11
idna
pydantic
python-dotenv
sniffio
SQLAlchemy
sqlalchemy[mypy]
starlette
typing_extensions
uvicorn[standard]
pydantic[email]
aiofiles
types-aiofiles
werkzeug
jinja2
python-jose[cryptography]
python-multipart
passlib[bcrypt]
types-passlib
pytest
pytest-asyncio
httpx
pytest-cov
